Better KScreen configuration management.
The previous method caused a SEGFAULT in libkscreen  when an external screen was toggled,
because the id changed and KScreen did not retrieve output in its list.
The strategy is now as follows: Refresh the configuration (especially output ids)
just before creating the new configuration. It works better but
sometimes causes an issue in plasmadesktop.

Also the calls the the config operation should use pointers,
which will be ay deleted in the next event loop using deleteLater()
in exec() method.